Update rtsc_end_coverage.py
Updated with annotation and so that the column name is FP_coverage and TP_coverage, depending on mode, rather than coverage for both. Also added an if statement, so that any transcripts that are shorter than the sum of -trim + _tp_l options are returned as NA. I've removed the print statement as all transcripts in the rtsc file are outputted.